+++ src/index.js
@@ -9,30 +9,15 @@ export default {
     const _io = fpm.app.io
     fpm.registerAction('BEFORE_SERVER_START', () => {
       _io.on( 'connection', ctx => {
-        fpm.publish('socketio.connection', {id: ctx.socket.id, data: ctx.data})
+        fpm.publish('socketio.connect', {id: ctx.socket.id})
       } )
       _io.on( 'message', ctx => {
-        _io.broadcast('message', ctx.data)
-        fpm.publish('socketio.message', ctx.data)
+        fpm.publish('socketio.message',  {id: ctx.socket.id, data: ctx.data})
       } )
-      _io.on( 'login', ctx => {
-        ctx.data.channel = 'online'
-        _io.broadcast('message', ctx.data)
-        fpm.publish('socketio.login', ctx.data)
-      } )
-    })
 
-    fpm.registerAction('BEFORE_MODULES_ADDED', (args) => {
-      let biz = args[0]
-      biz.m = _.assign(biz.m, {
-        websocket: {
-          broadcast: (arg) =>{
-            _io.broadcast('message', arg)
-            return {data: 1}
-          }
-        }
+      fpm.subscribe('socketio.broadcast', (topic, data) => {
+        _io.broadcast('message', data)
       })
-      
     })
   }
 }