+++ test/broadcast.test.js
@@ -1,7 +1,7 @@
-var should = require("should");
+var should = require("chai").should();
 var YF = require("yf-fpm-client-nodejs").default;
 
-YF.init({appkey: '123123', masterKey: '1b7e5703602b6fce1cae7364ac0f2244', endpoint: 'http://localhost:9999/api'});
+YF.init({appkey: '123123', masterKey: '123123', endpoint: 'http://localhost:9999/api'});
 
 
 describe('Websocket', function(){