+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.0.0 (2017-04-24)
+
+Feature
+- Add `http-server` Develope Dep, can RUN command: `yarn run dev:client` Then Test The websocket Client
+- You can Code `fpm.publish('socketio.broadcast', message)` To broadcast message
+
+
 ## 0.0.4 (2017-07-03)
 
 Feature